+++ CS_6610/Animation/Skeleton.cpp
@@ -51,7 +51,7 @@ void Skeleton::InitSimpleChain(Skeleton*& io_skeleton)
     io_skeleton->joints[0].local_to_parent.position_ = engine::math::Vec3D::ZERO;
 
     // Chain
-    for (uint16_t i = 1; i < io_skeleton->num_joints; ++i)
+    for (uint8_t i = 1; i < io_skeleton->num_joints; ++i)
     {
         io_skeleton->joints[i].local_to_parent.rotation_ = engine::math::Quaternion::FORWARD * engine::math::Quaternion::FORWARD;
         io_skeleton->joints[i].local_to_parent.position_.set(0.0f, io_skeleton->bone_length, 0.0f);
@@ -61,7 +61,7 @@ void Skeleton::InitSimpleChain(Skeleton*& io_skeleton)
 void Skeleton::InitHumanoid(Skeleton*& io_skeleton)
 {
     // Reset the transforms
-    for (uint16_t i = 0; i < io_skeleton->num_joints; ++i)
+    for (uint8_t i = 0; i < io_skeleton->num_joints; ++i)
     {
         io_skeleton->joints[i].local_to_parent.position_ = engine::math::Vec3D::ZERO;
         io_skeleton->joints[i].local_to_parent.rotation_ = engine::math::Quaternion::FORWARD * engine::math::Quaternion::FORWARD;
